Better error messages for api_version mismatch. Also be a bit stricter about when to set the error, and the types of version numbers.

+++ include/cwipc_util/api.h
@@ -23,13 +23,13 @@
 * Version of the current API of cwipc. Pass to constructors to ensure library
 * compatibility.
 */
-#define CWIPC_API_VERSION 0x20220328
+#define CWIPC_API_VERSION ((uint64_t)0x20220328)
 
 /** \brief Version of oldest compatible cwipc API.
 *
 * Version of the oldest API of cwipc to which this set of libraries is compatible.
 */
-#define CWIPC_API_VERSION_OLD 0x20211230
+#define CWIPC_API_VERSION_OLD ((uint64_t)0x20211230)
 
 /** \brief 4 characters that are magic number of cwipcdump file format
 */
@@ -37,7 +37,7 @@
 
 /** \brief Magic number (version) of cwipcdump file format
 */
-#define CWIPC_CWIPCDUMP_VERSION 0x20210208
+#define CWIPC_CWIPCDUMP_VERSION ((uint32_t)0x20210208)
 
 /** \brief flags for cwipc_write_ext. 1=write binary ply file
     */
